requirements.txt
gradio
Pillow
requests
httpx
google-genai
python-docx 
openpyxl
python-pptx



